Removed some unnecessary imports and corrected some style.

+++ backend/application/data_capture.py
@@ -1,4 +1,3 @@
-import numpy as np
 from cvzone.HandTrackingModule import HandDetector
 import cv2 as cv
 import sys
@@ -11,8 +10,6 @@ TODO: reference tutorial video
 
 
 def get_signs(filename):
-    desired_shape = (200, 200)
-    offset = 30
     hand_detector = HandDetector(maxHands=1)
     capture = cv.VideoCapture(0)
     i = 0
@@ -20,16 +17,12 @@ def get_signs(filename):
         s, img = capture.read()
         cv.imwrite(filename + str(i) + '.png', img)
         hands, hands_img = hand_detector.findHands(img)
-        try:
-            x, y, w, h = hands[0]['bbox']
-            cropped = hands_img[y-offset:y+offset+h, x-offset:x+w+offset]
-            cropped = cv.resize(cropped, desired_shape)
-            cv.imshow('img', img)
-            cv.waitKey(1)
-            i += 1
-        except Exception as e:
-            print(e)
+        if len(hands) < 1:
             os.remove(filename + str(i) + '.png')
+            continue
+        cv.imshow('img', img)
+        cv.waitKey(1)
+        i += 1
 
 
 if __name__ == '__main__':
